requirements.txt
flask==3.0.0
gunicorn==21.2.0
groq==0.4.1
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
